升级tkmapper 版本 user-provider 模块service 更新tkmapper

+++ user-service/user-provider/src/main/java/com/gpmall/user/dal/entitys/Address.java
@@ -25,6 +25,7 @@ public class Address {
     private String streetName;
 
     @Column(name = "is_default")
-    private Byte isDefault;
+    private Integer isDefault;
+
 
 }
